Moving xaml document files to View/Document
