--- CID_Tester/View/AddDutView1.xaml
@@ -1,26 +1,0 @@
-﻿<UserControl x:Class="CID_Tester.View.AddDutView1"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:CID_Tester.View"
-             xmlns:service="clr-namespace:CID_Tester.Service.Watermark"
-             mc:Ignorable="d" 
-             d:DesignHeight="450" d:DesignWidth="800">
-    <StackPanel Margin="20">
-        <TextBlock FontSize="18" Margin="0 0 0 20">Create New Device</TextBlock>
-        <TextBox Text="{Binding DutName, UpdateSourceTrigger=PropertyChanged}" Margin="0 0 0 10">
-            <service:WatermarkService.Watermark>
-                <TextBlock>Device name</TextBlock>
-            </service:WatermarkService.Watermark>
-        </TextBox>
-        <TextBox Text="{Binding DutDescription, UpdateSourceTrigger=PropertyChanged}" Margin="0 0 0 10" Height="150" AcceptsReturn="True" TextWrapping="Wrap" VerticalScrollBarVisibility="Auto">
-            <service:WatermarkService.Watermark>
-                <TextBlock>Description</TextBlock>
-            </service:WatermarkService.Watermark>
-        </TextBox>
-        <Button HorizontalAlignment="Left" Width="100" Command="{Binding AddDutCommand}">
-            <TextBlock Foreground="Black" Margin="1 2.5">Create Device</TextBlock>
-        </Button>
-    </StackPanel>
-</UserControl>
