--- CID_Tester/View/AddDutView1.xaml.cs
@@ -1,28 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
-
-namespace CID_Tester.View
-{
-    /// <summary>
-    /// Interaction logic for AddDutView1.xaml
-    /// </summary>
-    public partial class AddDutView1 : UserControl
-    {
-        public AddDutView1()
-        {
-            InitializeComponent();
-        }
-    }
-}
