--- CID_Tester/View/DashboardView.xaml.cs
@@ -1,29 +1,0 @@
-﻿using System.Diagnostics;
-using System.Windows;
-using System.Windows.Controls;
-
-namespace CID_Tester.View
-{
-    /// <summary>
-    /// Interaction logic for DashboardView.xaml
-    /// </summary>
-    public partial class DashboardView : UserControl
-    {
-        public DashboardView()
-        {
-            InitializeComponent();
-
-
-            //myDataGrid2.ItemsSource = context.TEST_PROCEDURE.ToList();
-            //myDataGrid3.ItemsSource = context.DUT.ToList();
-            //myDataGrid4.ItemsSource = context.TEST_USER.ToList();
-        }
-
-        private void Button_Click(object sender, RoutedEventArgs e)
-        {
-            var currentExecutablePath = Process.GetCurrentProcess().MainModule.FileName;
-            Process.Start(currentExecutablePath);
-            App.Current.Shutdown();
-        }
-    }
-}
