--- CID_Tester/View/DevicesView.xaml.cs
@@ -1,15 +1,0 @@
-﻿using System.Windows.Controls;
-
-namespace CID_Tester.View
-{
-    /// <summary>
-    /// Interaction logic for DevicesView.xaml
-    /// </summary>
-    public partial class DevicesView : UserControl
-    {
-        public DevicesView()
-        {
-            InitializeComponent();
-        }
-    }
-}
