+++ CID_Tester/View/Document/DashboardView.xaml
@@ -1,9 +1,9 @@
-﻿<UserControl x:Class="CID_Tester.View.DashboardView"
+﻿<UserControl x:Class="CID_Tester.View.Document.DashboardView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:CID_Tester.View" 
+             xmlns:local="clr-namespace:CID_Tester.View.Document"
              xmlns:control="clr-namespace:CID_Tester.View.Controls"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
@@ -13,7 +13,7 @@
                 <StackPanel Orientation="Horizontal" Margin="10">
                     <Ellipse Width="100" Height="100" Margin="20,0" Stroke="#1FB151" StrokeThickness="2">
                         <Ellipse.Fill>
-                            <ImageBrush ImageSource="../images/profile.jpg" />
+                            <ImageBrush ImageSource="../../images/profile.jpg" />
                         </Ellipse.Fill>
                     </Ellipse>
                     <StackPanel Orientation="Vertical">
