+++ CID_Tester/View/Document/DevicesView.xaml
@@ -1,12 +1,11 @@
-﻿<UserControl x:Class="CID_Tester.View.DevicesView"
+﻿<UserControl x:Class="CID_Tester.View.Document.DevicesView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:CID_Tester.View"
+             xmlns:local="clr-namespace:CID_Tester.View.Document"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
-
     <StackPanel Margin="10">
         <TextBlock Margin="0,0,0,10" FontSize="30">Devices Under Testing</TextBlock>
         <StackPanel Orientation="Horizontal" HorizontalAlignment="Right">
