+++ CID_Tester/View/Document/DevicesView.xaml.cs
@@ -1,0 +1,11 @@
+﻿using System.Windows.Controls;
+
+namespace CID_Tester.View.Document;
+
+public partial class DevicesView : UserControl
+{
+    public DevicesView()
+    {
+        InitializeComponent();
+    }
+}
