+++ CID_Tester/View/Document/ResultsView.xaml
@@ -1,13 +1,12 @@
-﻿<UserControl x:Class="CID_Tester.View.ResultsView"
+﻿<UserControl x:Class="CID_Tester.View.Document.ResultsView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:CID_Tester.View.Document"
              xmlns:wv2="clr-namespace:Microsoft.Web.WebView2.Wpf;assembly=Microsoft.Web.WebView2.Wpf"
-             xmlns:local="clr-namespace:CID_Tester.View"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
-
     <StackPanel Orientation="Vertical">
         <StackPanel Margin="20" HorizontalAlignment="Stretch" VerticalAlignment="Stretch" Orientation="Horizontal">
             <Border CornerRadius="7" Background="#2D2D30" Margin="0 0 20 0">
