+++ CID_Tester/View/Document/TestPlanView.xaml
@@ -1,9 +1,9 @@
-﻿<UserControl x:Class="CID_Tester.View.TestPlanView"
+﻿<UserControl x:Class="CID_Tester.View.Document.TestPlanView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:CID_Tester.View"
+             xmlns:local="clr-namespace:CID_Tester.View.Document"
              xmlns:controls="clr-namespace:CID_Tester.View.Controls"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
