+++ CID_Tester/View/Document/TestPlanView.xaml.cs
@@ -1,0 +1,11 @@
+﻿using System.Windows.Controls;
+
+namespace CID_Tester.View.Document;
+
+public partial class TestPlanView : UserControl
+{
+    public TestPlanView()
+    {
+        InitializeComponent();
+    }
+}
